Use `setup_test_component_platform` helper for text entity component tests instead of `hass.components`
