+++ tests/components/text/test_init.py
@@ -12,7 +12,6 @@ from homeassistant.components.text import (
     ATTR_VALUE,
     DOMAIN,
     SERVICE_SET_VALUE,
-    TextEntity,
     TextMode,
     _async_set_value,
 )
@@ -24,27 +23,9 @@ from homeassistant.setup import async_setup_component
 from tests.common import (
     async_mock_restore_state_shutdown_restart,
     mock_restore_cache_with_extra_data,
+    setup_test_component_platform,
 )
-
-
-class MockTextEntity(TextEntity):
-    """Mock text device to use in tests."""
-
-    def __init__(
-        self, native_value="test", native_min=None, native_max=None, pattern=None
-    ):
-        """Initialize mock text entity."""
-        self._attr_native_value = native_value
-        if native_min is not None:
-            self._attr_native_min = native_min
-        if native_max is not None:
-            self._attr_native_max = native_max
-        if pattern is not None:
-            self._attr_pattern = pattern
-
-    async def async_set_value(self, value: str) -> None:
-        """Set the value of the text."""
-        self._attr_native_value = value
+from tests.components.text.common import MockRestoreText, MockTextEntity
 
 
 async def test_text_default(hass: HomeAssistant) -> None:
@@ -126,21 +107,16 @@ RESTORE_DATA = {
 async def test_restore_number_save_state(
     hass: HomeAssistant,
     hass_storage: dict[str, Any],
-    enable_custom_integrations: None,
 ) -> None:
     """Test RestoreNumber."""
-    platform = getattr(hass.components, "test.text")
-    platform.init(empty=True)
-    platform.ENTITIES.append(
-        platform.MockRestoreText(
-            name="Test",
-            native_max=5,
-            native_min=1,
-            native_value="Hello",
-        )
+    entity0 = MockRestoreText(
+        name="Test",
+        native_max=5,
+        native_min=1,
+        native_value="Hello",
     )
+    setup_test_component_platform(hass, DOMAIN, [entity0])
 
-    entity0 = platform.ENTITIES[0]
     assert await async_setup_component(hass, "text", {"text": {"platform": "test"}})
     await hass.async_block_till_done()
 
@@ -167,7 +143,6 @@ async def test_restore_number_save_state(
 )
 async def test_restore_number_restore_state(
     hass: HomeAssistant,
-    enable_custom_integrations: None,
     hass_storage: dict[str, Any],
     native_max,
     native_min,
@@ -178,18 +153,14 @@ async def test_restore_number_restore_state(
     """Test RestoreNumber."""
     mock_restore_cache_with_extra_data(hass, ((State("text.test", ""), extra_data),))
 
-    platform = getattr(hass.components, "test.text")
-    platform.init(empty=True)
-    platform.ENTITIES.append(
-        platform.MockRestoreText(
-            native_max=native_max,
-            native_min=native_min,
-            name="Test",
-            native_value=None,
-        )
+    entity0 = MockRestoreText(
+        native_max=native_max,
+        native_min=native_min,
+        name="Test",
+        native_value=None,
     )
+    setup_test_component_platform(hass, DOMAIN, [entity0])
 
-    entity0 = platform.ENTITIES[0]
     assert await async_setup_component(hass, "text", {"text": {"platform": "test"}})
     await hass.async_block_till_done()
